+++ scripts/spells/darkness.js
@@ -1,13 +1,17 @@
 /*
-	You create a 20-foot-radius sphere of fog centered on a point within range. The sphere spreads around corners, and
-	its area is heavily obscured. It lasts for the duration or until a wind of moderate or greater speed (at least 10
-	miles per hour) disperses it.
+	Magical darkness spreads from a point you choose within range to fill a 15-foot-radius sphere for the duration. The
+	darkness spreads around corners. A creature with darkvision can't see through this darkness, and nonmagical light
+	can't illuminate it.
 
-	At Higher Levels. When you cast this spell using a spell slot of 2nd level or higher, the radius of the fog
-	increases by 20 feet for each slot level above 1st.
+	If the point you choose is on an object you are holding or one that isn't being worn or carried, the darkness
+	emanates from the object and moves with it. Completely covering the source of the darkness with an opaque object,
+	such as a bowl or a helm, blocks the darkness.
+
+	If any of this spell's area overlaps with an area of light created by a spell of 2nd level or lower, the spell that
+	created the light is dispelled.
 */
-const version = "11.0";
-const optionName = "Fog Cloud";
+const version = "11.1";
+const optionName = "Darkness";
 
 try {
 	if (args[0].macroPass === "templatePlaced") {
@@ -18,16 +22,16 @@ try {
 			let radius = canvas.grid.size * (templateDoc.distance / canvas.grid.grid.options.dimensions.distance);
 			await circleWall(templateDoc.x, templateDoc.y, radius);
 		}
-	}	
+	}
 	else if (args[0] === "off") {
 		async function removeWalls() {
-			let darkWalls = canvas.walls.placeables.filter(w => w.document.flags["midi-srd"]?.FogCloud?.ActorId === actor.id)
+			let darkWalls = canvas.walls.placeables.filter(w => w.document.flags["midi-srd"]?.Darkness?.ActorId === actor.id)
 			let wallArray = darkWalls.map(function (w) {
 				return w.document._id;
 			})
 			await canvas.scene.deleteEmbeddedDocuments("Wall", wallArray);
 		}
-		await removeWalls();		
+		await removeWalls();
 	}
 
 } catch (err) {
@@ -53,8 +57,9 @@ async function circleWall(cx, cy, radius) {
 			dir: CONST.WALL_DIRECTIONS.BOTH,
 			door: CONST.WALL_DOOR_TYPES.NONE,
 			ds: CONST.WALL_DOOR_STATES.CLOSED,
-			flags: { "midi-srd": { FogCloud: { ActorId: actor.id } } }
+			flags: { "midi-srd": { Darkness: { ActorId: actor.id } } }
 		});
 	}
+
 	await canvas.scene.createEmbeddedDocuments("Wall", data)
 }