+++ changelog.md
@@ -1,3 +1,14 @@
+# 11.13
+* Backgrounds Updated: all for dnd5e 2.4+
+* Classes Updated: all for dnd5e 2.4+
+* Creature Features Added: Deflect Missiles, Life Drain
+* Creatures Updated: Specter
+* Features Updated: Deflect Missiles (Monk), Sneak Attack
+* Macros Added: Draw Darkness Walls
+* Racial Traits Updated: Radiant Consumption
+* Spells Updated: Darkness, Fog Cloud
+* General: Fix for lack permission error when initiative is advanced
+
 # 11.12
 * Class Features Updated: Channel Divinity (Cleric), Channel Divinity (Paladin), Channel Divinity - Abjure Enemy, Channel Divinity - Turn Undead, Channel Divinity - Turn the Unholy, Divine Smite, Lay on Hands, Panache, Rage
 * Creatures Updated: Trolls (all), Zombie